v1.0.35 b 221
Added Scrolling Background
Changed Background Image
Fixed some coding issues
Fixed output spam
Added Speed multiplier while using "Beam" powerup
Fixed Sweeper-Detonator vs. Bomb Issues
Added refills on shield / beam

+++ src/glide/versioning/Updater.java
@@ -8,7 +8,9 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
+import java.net.HttpURLConnection;
 import java.net.URL;
+import java.net.UnknownHostException;
 import java.util.Random;
 
 public class Updater {
@@ -23,29 +25,33 @@ public class Updater {
 	}
 	
 	private void init() throws IOException, ClassNotFoundException{
-		String fname = rs(10) + ".v";
-		BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(fname, false));
-		BufferedInputStream in = new BufferedInputStream(online.openStream());
-	    byte data[] = new byte[1024];
-	    int count;
-	    while((count = in.read(data,0,1024)) != -1)
-	    {
-	        out.write(data, 0, count);
-	    }
-	    out.close();
-	    in.close();
-	    FileInputStream fin = new FileInputStream(fname);
-		ObjectInputStream ois = new ObjectInputStream(fin);
-		Version v = (Version) ois.readObject();
-		this.onlineVersion = v;
-		if(Version.getAppBuild() < this.onlineVersion.getBuild()){
-			this.needsUpdate = true;
-		}else{
-			this.needsUpdate = false;
+		if(internetCheck()){
+			String fname = rs(10) + ".v";
+			FileOutputStream fos = new FileOutputStream(fname, false);
+			BufferedOutputStream out = new BufferedOutputStream(fos);
+			BufferedInputStream in = new BufferedInputStream(online.openStream());
+			byte data[] = new byte[1024];
+			int count;
+			while((count = in.read(data,0,1024)) != -1)
+			{
+		        out.write(data, 0, count);
+		    }
+		    in.close();
+		    out.close();
+		    FileInputStream fin = new FileInputStream(fname);
+			ObjectInputStream ois = new ObjectInputStream(fin);
+			Version v = (Version) ois.readObject();
+			this.onlineVersion = v;
+			if(Version.getAppBuild() < this.onlineVersion.getBuild()){
+				this.needsUpdate = true;
+			}else{
+				this.needsUpdate = false;
+			}
+			fin.close();
+			ois.close();
+			fos.close();
+			new File(fname).delete();
 		}
-		fin.close();
-		ois.close();
-		new File(fname).delete();
 	}
 	
 	public boolean needsUpdate(){
@@ -66,4 +72,18 @@ public class Updater {
 	    }
 	    return ret;
 	}
+	public static boolean internetCheck()
+    {
+        try {
+            URL url = new URL("http://glide.fiscalleti.com/");
+            HttpURLConnection urlConnect = (HttpURLConnection)url.openConnection();
+            Object objData = urlConnect.getContent();
+        } catch (UnknownHostException e) {
+            return false;
+        }
+        catch (IOException e) {
+            return false;
+        }
+        return true;
+    }
 }
